+++ src/main/java/io/github/mathieusoysal/solution/SolutionsStorage.java
@@ -1,20 +1,20 @@
 package io.github.mathieusoysal.solution;
 
 class SolutionsStorage {
-    private long[] keys;
+    private int[] keys;
     private int size;
 
     public SolutionsStorage() {
-        keys = new long[2 << 15];
+        keys = new int[2 << 15];
         size = 0;
     }
 
     public SolutionsStorage(byte size) {
-        keys = new long[2 << size];
+        keys = new int[2 << size];
         size = 0;
     }
 
-    public boolean addSolution(long board, long nbUsesOfEachPieces) {
+    public boolean addSolution(int board, int nbUsesOfEachPieces) {
         int index = getIndex(board, nbUsesOfEachPieces);
         if (contains(board, nbUsesOfEachPieces)) {
             return false;
@@ -31,12 +31,12 @@ class SolutionsStorage {
         return true;
     }
 
-    public boolean contains(long board, long nbUsesOfEachPieces) {
+    public boolean contains(int board, int nbUsesOfEachPieces) {
         int index = getIndex(board, nbUsesOfEachPieces);
         return keys[index] == board && keys[index + 1] == nbUsesOfEachPieces;
     }
 
-    private int getIndex(long key1, long key2) {
+    private int getIndex(int key1, int key2) {
         int h1 = Long.hashCode(key1);
         int h2 = Long.hashCode(key2);
         return (h1 ^ h2) & (keys.length - 2);
@@ -53,12 +53,12 @@ class SolutionsStorage {
     private void resize() {
         long heapsize = Runtime.getRuntime().totalMemory();
         System.out.println("Heap size: " + heapsize * 0.000001 + " MB " + keys.length + " size");
-        long[] oldKeys = keys;
-        keys = new long[keys.length * 2];
+        int[] oldKeys = keys;
+        keys = new int[keys.length * 2];
         size = 0;
         for (int i = 0; i < oldKeys.length; i += 2) {
-            long key1 = oldKeys[i];
-            long key2 = oldKeys[i + 1];
+            int key1 = oldKeys[i];
+            int key2 = oldKeys[i + 1];
             if (key1 != 0 || key2 != 0) {
                 addSolution(key1, key2);
             }
